backend/requirements.txt
fastapi
uvicorn
pandas
pdfplumber
python-multipart
supabase
requests
python-dotenv
groq
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
